AssetMS/AssetManagement/requirements.txt
asgiref                          3.2.3
astroid                          2.3.3
colorama                         0.4.3
Django                           3.0.7
django-crispy-forms              1.8.1
django-filter                    2.2.0
django-login-required-middleware 0.4
django-widget-tweaks             1.4.5
django-wkhtmltopdf               3.4.0
imutils                          0.5.4
isort                            4.3.21
lazy-object-proxy                1.4.3
sqlparse                         0.3.0
tzdata                           2022.6
wkhtmltopdf                      0.2
wrapt                            1.11.2
xlwt                             1.3.0
asgiref                         3.5.2
attrs                           22.1.0
Django                          4.1.3
django-binary-database-files    1.0.17
django-pdf-generator            0.1.3
jsonschema                      4.17.0
mysql                           0.0.3
mysql-connector                 2.2.9
mysqlclient                     2.1.1
pip                             22.3.1
pyproject-toml                  0.0.10
pyrsistent                      0.19.2
setuptools                      63.2.0
sqlparse                        0.4.3
toml                            0.10.2
tzdata                          2022.6
wheel                           0.38.4
xlwt                            1.3.0